Update: Refactoring.
