+++ src/action/api.ts
@@ -1,16 +1,18 @@
 import { Ipc, BgCommand } from '@/services/ipc'
 import { sleep, toUrl, isValidString } from '@/lib/utils'
-import { ExecState } from './index'
-import type { ExecProps } from './index'
-import type { ApiCommand } from '@/types'
-import { OPEN_MODE } from '@/const'
+import type { ExecuteCommandParams, ApiCommand } from '@/types'
+import { OPEN_MODE, ExecState } from '@/const'
 
 const isApiCommand = (command: any): command is ApiCommand => {
   return command.OpenMode === OPEN_MODE.API
 }
 
 export const Api = {
-  async execute({ command, selectionText, changeState }: ExecProps) {
+  async execute({
+    command,
+    selectionText,
+    changeState = () => {},
+  }: ExecuteCommandParams) {
     changeState(ExecState.EXECUTING)
 
     if (!isValidString(command.searchUrl)) {
