+++ src/action/linkPreview.ts
@@ -6,11 +6,11 @@ import {
 } from '@/services/dom'
 import { getScreenSize } from '@/services/screen'
 import { DRAG_OPEN_MODE, POPUP_TYPE } from '@/const'
-import type { ExecProps } from './index'
 import { isEmpty } from '@/lib/utils'
+import type { ExecuteCommandParams } from '@/types'
 
 export const LinkPreview = {
-  async execute({ command, position, target }: ExecProps) {
+  async execute({ command, position, target }: ExecuteCommandParams) {
     if (position && target) {
       const elm = findAnchorElementFromParent(target) as HTMLAnchorElement
       const href = elm?.href
