+++ src/components/menu/MenuItem.tsx
@@ -8,8 +8,7 @@ import { linksInSelection } from '@/services/dom'
 import { useSelectContext } from '@/hooks/useSelectContext'
 import { useCommandExecutor } from '@/hooks/useCommandExecutor'
 import { sendEvent } from '@/services/analytics'
-import { OPEN_MODE } from '@/const'
-import { ExecState } from '@/action'
+import { OPEN_MODE, ExecState } from '@/const'
 import type { Command } from '@/types'
 
 import css from './Menu.module.css'
